requirements.txt
requests==2.31.0
pytest==7.4.4
pydantic==2.6.0
python-dotenv==1.0.1
allure-pytest==2.13.5